refactor: extract ComfyCliError and error-detail rendering into errors.py
Moves three clusters from server.py into a new leaf module,
src/comfy_mcp/errors.py, byte-for-byte: the ComfyCliError exception
(server.py's one comfy-cli failure type), the benign "nothing recorded
to stop" detector restart_comfyui relies on (_is_no_recorded_server and
its two marker regexes), and the error.details renderer plus the
per-field char cap every error message in the module is bounded by
(_render_error_details, _SURFACED_DETAIL_KEYS, _MAX_ERROR_FIELD_CHARS).

server.py reaches the private names module-qualified (errors._X) at
every call site, including the ~25 pre-existing uses of
_MAX_ERROR_FIELD_CHARS scattered through code that stays in server.py
— both sides now read errors._MAX_ERROR_FIELD_CHARS at call time rather
than an import-bound copy, so a test patching it still reaches the one
copy that matters. One collision fell out of the mechanical rename:
_local_template_check had an unrelated local variable also named
`errors` (the workflow's own validation errors), which would have
shadowed the module import; renamed to `validation_errors` there.

ComfyCliError itself is the carve-out AGENTS.md now documents: a
public exception type is name-imported (`from .errors import
ComfyCliError`) rather than reached module-qualified, since it rides
hundreds of `except`/`isinstance` sites and has no mutable state a test
could silently patch the wrong copy of. test_failure_log,
test_partner_auth, test_workflow, and test_wrapper rename their
_MAX_ERROR_FIELD_CHARS / _is_no_recorded_server / _NO_RECORDED_SERVER_CODE
references (19 total) to the errors module; none needed anything more.
AGENTS.md's module table gains an errors.py row and its re-export
sentence is amended once for the type carve-out.
